1、分页分库增强 2、INYINT类型的转为JAVA的Long类型

+++ src/main/java/com/raycloud/util/daogen/SqlType2Feild.java
@@ -23,7 +23,8 @@ public class SqlType2Feild {
 		sqltype2Feild.put(Types.DATE, "Date");	//java.util.Date
 		//mysql
 		sqltype2Feild.put(Types.BIT, "Byte");
-		sqltype2Feild.put(Types.TINYINT, "Byte");
+//		sqltype2Feild.put(Types.TINYINT, "Byte");
+        sqltype2Feild.put(Types.TINYINT, "Long");
 		sqltype2Feild.put(Types.SMALLINT, "Integer");
 		sqltype2Feild.put(Types.INTEGER, "Integer");
 		sqltype2Feild.put(Types.BIGINT, "Long");
